claude: update code based on gh comment feed back
